requirements.txt
Django>=5.1,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.2
